[+] add logstash logger [-] remove fluent logger
